Reorganized sass files to be in src/styles instead of src/assets/css
