+++ src/components/metadataEditor/metadataEditor.js
@@ -15,8 +15,8 @@ import '../../elements/emby-textarea/emby-textarea';
 import '../../elements/emby-button/emby-button';
 import '../../elements/emby-button/paper-icon-button-light';
 import '../formdialog.scss';
-import '../../assets/css/clearbutton.scss';
-import '../../assets/css/flexstyles.scss';
+import '../../styles/clearbutton.scss';
+import '../../styles/flexstyles.scss';
 import './style.scss';
 import ServerConnections from '../ServerConnections';
 import toast from '../toast/toast';
