+++ src/components/sortmenu/sortmenu.js
@@ -7,7 +7,7 @@ import '../../elements/emby-button/paper-icon-button-light';
 import 'material-design-icons-iconfont';
 import '../formdialog.scss';
 import '../../elements/emby-button/emby-button';
-import '../../assets/css/flexstyles.scss';
+import '../../styles/flexstyles.scss';
 import template from './sortmenu.template.html';
 
 function onSubmit(e) {
