+++ src/controllers/livetvstatus.js
@@ -7,7 +7,7 @@ import layoutManager from '../components/layoutManager';
 import loading from '../components/loading/loading';
 import browser from '../scripts/browser';
 import '../components/listview/listview.scss';
-import '../assets/css/flexstyles.scss';
+import '../styles/flexstyles.scss';
 import '../elements/emby-itemscontainer/emby-itemscontainer';
 import '../components/cardbuilder/card.scss';
 import 'material-design-icons-iconfont';
