+++ src/controllers/wizard/user/index.js
@@ -1,6 +1,6 @@
 import loading from '../../../components/loading/loading';
 import globalize from '../../../scripts/globalize';
-import '../../../assets/css/dashboard.scss';
+import '../../../styles/dashboard.scss';
 import '../../../elements/emby-input/emby-input';
 import '../../../elements/emby-button/emby-button';
 import Dashboard from '../../../utils/dashboard';
