+++ src/libraries/scroller.js
@@ -8,7 +8,7 @@ import layoutManager from '../components/layoutManager';
 import dom from '../scripts/dom';
 import focusManager from '../components/focusManager';
 import ResizeObserver from 'resize-observer-polyfill';
-import '../assets/css/scrollstyles.scss';
+import '../styles/scrollstyles.scss';
 import globalize from '../scripts/globalize';
 
 /**
