+++ src/scripts/libraryMenu.js
@@ -22,8 +22,8 @@ import { getParameterByName } from '../utils/url.ts';
 import '../elements/emby-button/paper-icon-button-light';
 
 import 'material-design-icons-iconfont';
-import '../assets/css/scrollstyles.scss';
-import '../assets/css/flexstyles.scss';
+import '../styles/scrollstyles.scss';
+import '../styles/flexstyles.scss';
 
 /* eslint-disable indent */
