+++ src/scripts/routes.js
@@ -7,10 +7,10 @@ import '../elements/emby-collapse/emby-collapse';
 import '../elements/emby-select/emby-select';
 import '../elements/emby-checkbox/emby-checkbox';
 import '../elements/emby-slider/emby-slider';
-import '../assets/css/livetv.scss';
+import '../styles/livetv.scss';
 import '../components/listview/listview.scss';
-import '../assets/css/dashboard.scss';
-import '../assets/css/detailtable.scss';
+import '../styles/dashboard.scss';
+import '../styles/detailtable.scss';
 import { appRouter } from '../components/appRouter';
 
 /* eslint-disable indent */
