+++ src/styles/clearbutton.scss
