+++ src/styles/detailtable.scss
