+++ src/styles/librarybrowser.scss
