+++ src/styles/metadataeditor.scss
